Rename OMSUtil to OMSClientUtil
